AppendableFile goes from filesnpaths to utils.files
